feat: revamp queue management forms and enhance localization
- Introduced new Queue and QueueInputsForm components to streamline queue management within the settings.
- Updated localization files for English, Spanish, and Portuguese to include new keys for queue-related functionalities, improving user guidance.
- Refactored existing components to support the new queue structure, ensuring a cohesive user experience.
- Enhanced the ListSectorQueues component to accommodate the new queue management logic, improving overall functionality.
- Removed deprecated Queue.vue component to clean up the codebase.
