+++ src/locales/pt_br.json
@@ -943,11 +943,15 @@
     },
     "queues": {
       "in_sector": "Fila no setor {sector}",
-      "delete_or_edit": "Excluir ou editar",
       "new": "Nova fila",
       "info": "As filas são grupos de atendimento dentro de um setor, é possível direcionar contatos de um fluxo para uma fila específica",
       "hint": "As filas são grupos de atendimento dentro do setor, você poderá criar novas filas ou editar quando quiser",
-      "queue_definitions": "Definições da fila",
+      "add_queue": "Adicionar fila",
+      "configure_queue": {
+        "title": "Configurar fila",
+        "tooltip": "As filas são grupos de atendimento dentro do setor. Você pode criar novas filas ou editá-las quando quiser após criar o setor."
+      },
+      "delete_or_edit": "Excluir ou editar",
       "limit_chats": {
         "title": "Limite de chats aguardando suporte",
         "switch": {
